+++ R/addVarEEclosestImage.R
@@ -23,7 +23,10 @@
 #' computing means, counts, etc. Sometimes we might want to avoid this behaviour
 #' and use 0 instead of NA. If so, set unmask to TRUE.
 #'
-#' @return
+#' @return A dataframe similar to \code{ee_pentads} with variables added from the
+#' \code{bands} selected from \code{collection}. Note that following \href{https://github.com/r-spatial/rgee}{rgee}
+#' the name of the new variables will be the selected band (\code{bands} or else
+#' all bands from \code{collection} followed by the spatial reducer \code{reducer}.
 #' @export
 #'
 #' @examples
@@ -43,7 +46,7 @@ addVarEEclosestImage <- function(ee_pentads, collection, reducer, maxdiff,
 
   # Get image
   if(is.character(collection)){
-    ee_layer <- ee$ImageCollection(collection)
+    ee_layer <- rgee::ee$ImageCollection(collection)
   } else if("ee.imagecollection.ImageCollection" %in% class(collection)){
     ee_layer <- collection
   } else {
@@ -66,7 +69,7 @@ addVarEEclosestImage <- function(ee_pentads, collection, reducer, maxdiff,
 
   # Function to add date in milliseconds
   addTime <- function(feature) {
-    datemillis <- ee$Date(feature$get('Date'))$millis()
+    datemillis <- rgee::ee$Date(feature$get('Date'))$millis()
     return(feature$set(list('date_millis' = datemillis)))
   }
 
@@ -74,14 +77,14 @@ addVarEEclosestImage <- function(ee_pentads, collection, reducer, maxdiff,
   ee_pentads <- ee_pentads$map(addTime)
 
   # Set filter to select images within max time difference
-  maxDiffFilter = ee$Filter$maxDifference(
+  maxDiffFilter = rgee::ee$Filter$maxDifference(
     difference = maxdiff*24*60*60*1000,        # days * hr * min * sec * milliseconds
     leftField = "date_millis",                 # Timestamp of the visit
     rightField = "system:time_start"           # Image date
   )
 
   # Set a saveBest join that finds the image closest in time
-  saveBestJoin <- ee$Join$saveBest(
+  saveBestJoin <- rgee::ee$Join$saveBest(
     matchKey = "bestImage",
     measureKey = "timeDiff"
   )
@@ -90,16 +93,16 @@ addVarEEclosestImage <- function(ee_pentads, collection, reducer, maxdiff,
   best_matches <- saveBestJoin$apply(ee_pentads, ee_layer, maxDiffFilter)
 
   # Function to add value from the matched image
-  reducer <- paste0("ee$Reducer$", reducer, "()")
+  reducer <- paste0("rgee::ee$Reducer$", reducer, "()")
   add_value <- function(feature){
 
     # Get the image selected by the join
-    img <- ee$Image(feature$get("bestImage"))$select(bands)
+    img <- rgee::ee$Image(feature$get("bestImage"))$select(bands)
 
     # Reduce values within pentad
     pentad_val <- img$reduceRegion(eval(parse(text = reducer)),
-                                  feature$geometry(),
-                                  scale)
+                                   feature$geometry(),
+                                   scale)
 
     # Return the data containing value and image date.
     return(feature$set('val', pentad_val$get(bands),
@@ -109,7 +112,15 @@ addVarEEclosestImage <- function(ee_pentads, collection, reducer, maxdiff,
 
   # Add values to the data and download
   out <- best_matches$map(add_value) %>%
-    ee_as_sf(via = 'drive')
+    rgee::ee_as_sf(via = 'drive')
+
+  # Fix names and variables
+  layer_name <- paste0(bands, "_", reducer)
+
+  out <- out %>%
+      dplyr::rename_with(~gsub("val", layer_name, .x), .cols = dplyr::starts_with("val")) %>%
+      dplyr::select(-c(id, date_millis, bestImage)) %>%
+      dplyr::select(dplyr::everything(), DateTimeImage)
 
   return(out)