+++ NEWS.md
@@ -1,3 +1,19 @@
-# ABAP 0.0.4
+# ABAP 0.1.1
+
+* Speed up 'abapTo' functions
+
+* spOccupancy and unmarked objects are ordered by pentad
+
+# ABAP 0.1.0
+
+* Set a standard for variables coming out of Google Earth Engine images. Variables
+will always be named as \"\<bandname\>_\<spt_reducer\>\". **Note that this change
+could break some existing code**
+
+* Added functions to format ABAP data into objects used by popular occupancy modelling
+R packages, such as *unmarked* or *spOccupancy*.
+
+# ABAP 0.0.5
 
 * Added a `NEWS.md` file to track changes to the package.
+* rgee:: called explicitly